sham algo trade/requirements.txt
kiteconnect==4.2.0
python-dotenv==1.0.0
schedule==1.2.0
twilio==8.10.0
requests==2.31.0
pyotp==2.9.0
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
numpy==1.26.2
python-telegram-bot==20.7
cryptography==41.0.7
sqlalchemy==2.0.23
flask==3.0.0
flask-cors==4.0.0
